backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.9
python-decouple==3.8
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0 